новая разбивка по модулям
